Dirac Operators

Dirac operators are structs that hold a reference to the gauge background U, a temporary fermion field in case one wants to use the doubly flavoured Hermitian variant, the bare mass and a Boolean indicating whether there are antiperiodic boundary conditions in the time direction (yes, only periodic and antiperiodic BCs are supported so far).

To create a Dirac operator, the constructors below are used. One thing to note is that dirac operators can be constructed using any Abstractfield and so the gauge background is always set to nothing on construction. In order to then add a gauge background you must use the dirac operator as a functor on a Gaugefield, like D_U = D_free(U). This does not overwrite the U in D_free but creates a new dirac operator, that references the same temporary fermion fields as the parent and does therefore not introduce any new allocations of fields.

MetaQCD.DiracOperators.WilsonDiracOperator — Type

WilsonDiracOperator(::Abstractfield, mass; anti_periodic=true, r=1, csw=0)
WilsonDiracOperator(D::WilsonDiracOperator, U::Gaugefield)

Create a free Wilson Dirac Operator with mass mass and Wilson parameter r. If anti_periodic is true the fermion fields are anti periodic in the time direction. If csw ≠ 0, a clover term is included. This object cannot be applied to a fermion vector, since it lacks a gauge background. A Wilson Dirac operator with gauge background is created by applying it to a Gaugefield U like D_gauge = D_free(U)

Type Parameters:

B: Backend (CPU / CUDA / ROCm)

T: Floating point precision

TF: Type of the Fermionfield used to store intermediate results when using the Hermitian version of the operator

TG: Type of the underlying Gaugefield

C: Boolean declaring whether the operator is clover improved or not

MetaQCD.DiracOperators.StaggeredDiracOperator — Type

StaggeredDiracOperator(::Abstractfield, mass; anti_periodic=true)
StaggeredDiracOperator(D::StaggeredDiracOperator, U::Gaugefield)

Create a free Staggered Dirac Operator with mass mass. If anti_periodic is true the fermion fields are anti periodic in the time direction. This object cannot be applied to a fermion vector, since it lacks a gauge background. A Staggered Dirac operator with gauge background is created by applying it to a Gaugefield U like D_gauge = D_free(U)

Type Parameters:

B: Backend (CPU / CUDA / ROCm)

T: Floating point precision

TF: Type of the Fermionfield used to store intermediate results when using the Hermitian version of the operator

TG: Type of the underlying Gaugefield

MetaQCD.DiracOperators.StaggeredEOPreDiracOperator — Type

StaggeredEOPreDiracOperator(::Abstractfield, mass; anti_periodic=true)
StaggeredEOPreDiracOperator(
    D::Union{StaggeredDiracOperator,StaggeredEOPreDiracOperator},
    U::Gaugefield
)

Create a free even-odd preconditioned Staggered Dirac Operator with mass mass. If anti_periodic is true the fermion fields are anti periodic in the time direction. This object cannot be applied to a fermion vector, since it lacks a gauge background. A Staggered Dirac operator with gauge background is created by applying it to a Gaugefield U like D_gauge = D_free(U)

Type Parameters:

B: Backend (CPU / CUDA / ROCm)

T: Floating point precision

TF: Type of the Fermionfield used to store intermediate results when using the Hermitian version of the operator

TG: Type of the underlying Gaugefield