Biased Sampling Methods

MetaQCD.BiasModule.calc_weights — Method

Weighting schemes based on the ones compared in
https://pubs.acs.org/doi/pdf/10.1021/acs.jctc.9b00867

MetaQCD.BiasModule.Bias — Type

Bias(p::ParameterSet, U::Gaugefield; instance=1)

Container that holds general parameters of bias enhanced sampling, like the kind of CV, its smearing and filenames/-pointers relevant to the bias. Also holds the specific kind of bias (Metadynamics, OPES or Parametric for now).
The instance keyword is used in case of PT-MetaD and multiple walkers to assign the correct usebias to each stream.

MetaQCD.BiasModule.Metadynamics — Type

Metadynamics(; symmetric=true, stride=1, cvlims=(-6, 6), biasfactor=Inf,
              bin_width=0.1, weight=0.01, penalty_weight=1000)
Metadynamics(p::ParameterSet; instance=1)

Create an instance of a Metadynamics bias using the inputs or the parameters given in p.

Specifiable parameters

symmetric::Bool = true - If true, the bias is built symmetrically by updating for both cv and -cv at every update-iteration
stride::Int64 = 1 - Number of iterations between updates; must be >0
cvlims::NTuple{2, Float64} = (-6, 6) - Minimum and maximum of the explorable cv-space; must be ordered
biasfactor::Float64 = Inf - Biasfactor for well-tempered Metadynamics; must be >1
bin_width::Float64 = 0.1 - Width of bins in histogram; must be >0
weight::Float64 = 0.01 - (Starting) Height of added Gaussians; must be positive
penalty_weight::Float64 = 1000 - Penalty when cv is outside of cvlims; must be positive

MetaQCD.BiasModule.OPES — Type

OPES(; symmetric=true, stride=1, cvlims=(-6, 6), barrier=30,
     biasfactor=Inf, σ₀=0.1, σ_min=1e-6, fixed_σ=true, opes_epsilon=0.0,
     no_Z=false, threshold=1.0, cutoff=0.0)
OPES(p::ParameterSet; instance=1)

Create an instance of a OPES bias using the parameters given in p.

Specifiable parameters

symmetric::Bool = true - If true, the bias is built symmetrically by updating for both cv and -cv at every update-iteration
stride::Int64 = 1 - Number of iterations between updates; must be >0
cvlims::NTuple{2, Float64} = (-6, 6) - Minimum and maximum of the explorable cv-space; must be ordered
barrier::Float64 = 30 - Estimate of height of action barriers
biasfactor::Float64 = Inf - Biasfactor for well-tempered OPES; must be >1
σ₀::Float64 = 0.1 - (Starting) width of kernels; must be >0
σ_min::Float64 = 1e-6 - Minimum width of kernels; must be >0
fixed_σ::Bool = true - If true, width if kernels decreases iteratively
ϵ::Float64 = exp(-barrier/(1-1/biasfactor)) - Determines maximum height of bias; must be >0
no_Z::Bool = false - If false normalization factor Z is dynamically adjusted
threshold::Float64 = 1.0 - Threshold distance for kernel merging; must be >0
cutoff::Float64 = sqrt(2barrier/(1-1/biasfactor)) - Cutoff value for kernels; must be >0
penalty::Float64 = exp(-0.5cutoff²) - Penalty for being outside kernel cutoff; must be >0

MetaQCD.BiasModule.Parametric — Type

Parametric(cvlims, penalty_weight, Q, A, Z)
Parametric(p::ParameterSet; instance=1)

Create an instance of a static Parametric bias using the inputs or the parameters given in p.

Specifiable parameters

cvlims::NTuple{2, Float64} = (-6, 6) - Minimum and maximum of the explorable cv-space; must be ordered
penalty_weight::Float64 = 1000 - Penalty when cv is outside of cvlims; must be positive
Q::Float64 = 0 - Quadratic term in the bias
A::Float64 = 0 - Amplitude of the cosine term in the bias
Z::Float64 = 0 - Frequency of the cosine term in the bias