Creating Fields on the lattice

You can create a 4-dimensional SU(3) gauge field by specifying the backend, floating point precision, gauge action, dimensions and coupling parameter beta:

backend = CPU
prec = Float64
action = WilsonGaugeAction
Ns = Nt = 12
beta = 6.0
U = Gaugefield{backend,prec,action}(Ns, Ns, Ns, Nt, beta) # all links are set to 0

and set the initial conditions with identity_gauges!(U) (cold) or random_gauges!(U) (hot).

Gaugefields, Colorfields and Expfields are structs that contain a main Array U, which is a 5-dimensional array of statically sized 3x3 complex matrices, i.e., SMatrix objects from StaticArrays.jl (where arrays are stored as Tuples under the hood).

The fact that the elements are statically sized immutable arrays means that, for one, there are no allocations when performing linear algebra operations with them and secondly that we always just override the matrices in the arrays instead of mutating them. This yields enormous benefits in terms of less headaches during development and lets us define custom linear algebra routines for SMatrices and SVectors.

The different backends are handled by Kernelabstractions.jl.

There are no plans to use more memory efficient storage schemes for SU(3) or su(3) elements yet.

Fermion fields or spinors or whatever you want to call them are stored in 4-dimensional arrays of n_color * n_dirac complex valued SVectors. The reason for chosing 4 instead of 5 dimensions is that this enabled us to write routines that take care of all dirac components at the same time, which is more efficient (I think).

When using even-odd preconditioned dirac operators, the fermion fields get wrapped in a struct called EvenOdd such that we can overload all functions on that type. Our convention is to define the fields on the even sites. While we haven't tested whether the following is actually more performant, we map all even sites to the first half of the array to have contiguous memory accesses. The function eo_site does exactly this mapping.

Fermionfields are created in the same way as Gaugefields with the gauge action type parameter being replaced by the number of Dirac indices. For Fermionfields we have the ones! and gaussian_pseudofermions! methods to init them.

MetaQCD.Fields.Gaugefield — Type

Gaugefield{BACKEND,T,GA}(NX, NY, NZ, NT, β)
Gaugefield(U::Gaugefield)

Creates a Gaugefield on BACKEND, i.e. an array of link-variables (SU3 matrices with T precision) of size 4 × NX × NY × NZ × NT with coupling parameter β and gauge action GA or a zero-initialized copy of U

Supported backends

CPU
CUDABackend
ROCBackend

Supported gauge actions

WilsonGaugeAction
SymanzikTreeGaugeAction (Lüscher-Weisz)
IwasakiGaugeAction
DBW2GaugeAction

MetaQCD.Fields.Fermionfield — Type

Fermionfield{BACKEND,T,ND}(NX, NY, NZ, NT)
Fermionfield(ψ::Fermionfield)
Fermionfield(f::Abstractfield, staggered)

Creates a Fermionfield on BACKEND, i.e. an array of link-variables (NC×ND complex vectors with T precision) of size NX × NY × NZ × NT or a zero-initialized copy of ψ. If staggered=true, the number of Dirac degrees of freedom (ND) is reduced to 1 instead of 4.

Supported backends

CPU
CUDABackend
ROCBackend

MetaQCD.Fields.Colorfield — Type

Colorfield{BACKEND,T}(NX, NY, NZ, NT)
Colorfield(u::Abstractfield)

Creates a Colorfield on BACKEND, i.e. an array of 3-by-3 T-precision matrices of size 4 × NX × NY × NZ × NT or a zero-initialized Colorfield of the same size as u

Supported backends

CPU
CUDABackend
ROCBackend

MetaQCD.Fields.Expfield — Type

Expfield{BACKEND,T}(NX, NY, NZ, NT)
Expfield(u::Abstractfield)

Creates a Expfield on BACKEND, i.e. an array of T-precison exp_iQ_su3 objects of size 4 × NX × NY × NZ × NT or of the same size as u. The objects hold the Q-matrices and all the exponential parameters needed for stout-force recursion

Supported backends

CPU
CUDABackend
ROCBackend